+++ packages/tdesign-uniapp-chat/CHANGELOG.md
@@ -5,6 +5,10 @@ toc: false
 docClass: timeline
 ---
 
+## 🌈 0.1.2 `2025-11-25` 
+### 🚀 Features
+- `ChatMarkdown`: 兼容 `Vue2.x` @novlan1 ([#92](https://github.com/novlan1/tdesign-uniapp/pull/92))
+
 ## 🌈 0.1.1 `2025-11-17` 
 ### 🐞 Bug Fixes
 - `Attachments`: 修复图标点击事件冒泡问题 @novlan1 ([#72](https://github.com/novlan1/tdesign-uniapp/pull/72))